web_app/requirements.txt
# Backend Dependencies
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.1
bcrypt==4.1.2

# Frontend Dependencies (will be managed through package.json)
# Using React + TypeScript

# Testing and Development
pytest==8.0.2
black==24.2.0
flake8==7.0.0 